+++ lib/Traits/TraitsClasses/GeoJsonTraits.ts
@@ -5,7 +5,6 @@ import objectTrait from "../Decorators/objectTrait";
 import primitiveTrait from "../Decorators/primitiveTrait";
 import mixTraits from "../mixTraits";
 import ModelTraits from "../ModelTraits";
-import FeatureInfoTraits from "./FeatureInfoTraits";
 import LegendOwnerTraits from "./LegendOwnerTraits";
 import StyleTraits from "./StyleTraits";
 import TableTraits from "./TableTraits";
@@ -39,7 +38,6 @@ export class PerPropertyGeoJsonStyleTraits extends ModelTraits {
 export class GeoJsonTraits extends mixTraits(
   LegendOwnerTraits,
   TableTraits,
-  FeatureInfoTraits,
   UrlTraits
 ) {
   /** Override TableTraits which aren't applicable to GeoJsonTraits */