+++ lib/Traits/TraitsClasses/MappableTraits.ts
@@ -3,6 +3,7 @@ import primitiveTrait from "../Decorators/primitiveTrait";
 import mixTraits from "../mixTraits";
 import ModelTraits from "../ModelTraits";
 import AttributionTraits from "./AttributionTraits";
+import { FeatureInfoTemplateTraits } from "./FeatureInfoTraits";
 
 export class RectangleTraits extends ModelTraits {
   @primitiveTrait({
@@ -253,4 +254,20 @@ export default class MappableTraits extends mixTraits(AttributionTraits) {
       "A message to show when the user adds the catalog item to the workbench. Useful for showing disclaimers."
   })
   initialMessage?: InitialMessageTraits;
+
+  @objectTrait({
+    type: FeatureInfoTemplateTraits,
+    name: "Feature info template",
+    description:
+      "A template object for formatting content in feature info panel"
+  })
+  featureInfoTemplate?: FeatureInfoTemplateTraits;
+
+  @primitiveTrait({
+    type: "string",
+    name: "Show string if property value is null",
+    description:
+      "If the value of a property is null or undefined, show the specified string as the value of the property. Otherwise, the property name will not be listed at all."
+  })
+  showStringIfPropertyValueIsNull?: string;
 }