+++ lib/Traits/TraitsClasses/SdmxCatalogItemTraits.ts
@@ -2,7 +2,6 @@ import objectArrayTrait from "../Decorators/objectArrayTrait";
 import primitiveTrait from "../Decorators/primitiveTrait";
 import mixTraits from "../mixTraits";
 import EnumDimensionTraits from "./DimensionTraits";
-import FeatureInfoTraits from "./FeatureInfoTraits";
 import LegendOwnerTraits from "./LegendOwnerTraits";
 import SdmxCommonTraits from "./SdmxCommonTraits";
 import TableTraits from "./TableTraits";
@@ -20,7 +19,6 @@ export class SdmxDimensionTraits extends mixTraits(EnumDimensionTraits) {
 export default class SdmxCatalogItemTraits extends mixTraits(
   SdmxCommonTraits,
   UrlTraits,
-  FeatureInfoTraits,
   TableTraits,
   LegendOwnerTraits
 ) {